[feature](amdin):user,department,batch module added

+++ client/js/dashboard.js
@@ -8,12 +8,11 @@ document.addEventListener("DOMContentLoaded", () => {
   const parseToken = JSON.parse(decodeToken);
   const role = parseToken?.type;
 
-  //   console.log(parseUser, role, token, decodeToken, parseToken);
-  if (parseUser.role !== role && token && role) {
+  console.log(parseUser, role, token, decodeToken, parseToken);
+  if (parseUser.role !== role && token && role && role !== "admin") {
     localStorage.clear();
     window.location.href = "index.html";
   }
-
   //   if (!token || !role) {
   //     window.location.href = "index.html";
   //   } else {
@@ -21,10 +20,9 @@ document.addEventListener("DOMContentLoaded", () => {
   //       "welcomeMessage"
   //     ).innerText = `Logged in as Role: ${role}`;
   //   }
-
   // Logout Functionality
-  document.getElementById("logoutBtn").addEventListener("click", () => {
-    localStorage.clear();
-    window.location.href = "index.html";
-  });
+  // document.getElementById("logoutBtn").addEventListener("click", () => {
+  //   localStorage.clear();
+  //   window.location.href = "index.html";
+  // });
 });
